+++ app/src/main/java/com/texthip/thip/data/model/users/response/MyFollowingsResponse.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.data.model.users
+package com.texthip.thip.data.model.users.response
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
