+++ app/src/main/java/com/texthip/thip/data/model/users/response/NicknameResponse.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.data.model.users
+package com.texthip.thip.data.model.users.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
